Draft Python Notebook with Kaggle Data
This is a draft notebook of what I have done so far with JUST the Kaggle data, not joined with any other dataset. Feel free to use this code with the joined dataset. There could be more algorithms that can be used, so feel free to download and add more algorithms as well :)